refactor: #122 split default layout into components
